more `getItems` and `getFields` parsers into a shared util
In `getItems`, we were skipping rows because blank lines in a tabular file were being skipping as comment lines, and therefore we had count mismatch between the total rows and comment rows.

This was happening in duplicated spots, so created a shared util.
